completed till register

+++ src/main/java/FinanceManangementSystem/demo/Repository/UserAddressRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAddressRepository extends JpaRepository<UserAddress,Integer> {
+public interface UserAddressRepository extends JpaRepository<UserAddress,Long> {
 }
